pyproject.toml
[tool.poetry]
name = "test-for-mera-capital"
version = "0.1.0"
description = ""
authors = ["Dmitry Spiryaev imwisagist@yandex.ru"]
readme = "README.md"
packages = [{include = "test_for_mera_capital"}]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.4"
apscheduler = "^3.10.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.16"}
pytest-asyncio = "^0.21.0"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
fastapi = "^0.95.0"
uvicorn = {extras = ["standart"], version = "^0.21.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
